Extract 'create' command into separate file
I am pretty much sold on 'yargs' as my argument parser of choice. So far
so good in extracting the logic of a more complex command out into a
separate file. TypeScript is a godsend in that it helps with type
checking in such an expressive language. Never going back.